requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
decorator==5.1.1
Flask==3.0.1
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
user-agents-parser==2.2.3